+++ Sources/MapboxNavigationCore/Map/Style/RouteAlertsAnnotationsMapFeatures.swift
@@ -11,8 +11,9 @@ struct RouteAlertsStyleContent: MapStyleContent {
 
     var body: some MapStyleContent {
         source
+
         symbolLayer
-            .slot(nil)
+            .slot(NavigationSlot.aboveBasemap)
     }
 }